backend/requirements.txt
# ════════════════════════════════════════════════════════════════
# backend/requirements.txt
# ════════════════════════════════════════════════════════════════
Django==5.0.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
django-filter==24.*
django-celery-beat==2.6.*
psycopg2-binary==2.9.*
redis==5.0.*
celery==5.4.*

# Scraping
scrapy==2.11.*
scrapy-playwright==0.0.34
playwright==1.44.*
beautifulsoup4==4.12.*
lxml==5.2.*
itemadapter==0.9.*
fake-useragent==1.5.*
python-dateutil==2.9.*

# Utilities
python-dotenv==1.0.*
Pillow==10.*
gunicorn==22.*
